+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Remove unused `getTimetableImage` function in `Export/GetImages.hs`
 - Refactored various backend text functions and tests to avoid `String` data in favour of `Text` when feasible
 - Removed unused files
+- Refactored `returnPDF`, `exportTimetablePDFResponse`, and `graphImageResponse` to use stdin instead of temporary `.tex` and `.svg` files
 
 ## [0.7.2] - 2025-12-10